requirements.txt
torch>=2.0.0
torchvision>=0.15.0
tqdm>=4.0.0
pillow>=10.0.0
requests>=2.31.0
geopy>=2.3.0
scipy>=1.12.0
numpy>=1.26.0
shapely>=2.0.0
pyproj>=3.6.1
fiona>=1.9.5
rtree>=1.2.0
matplotlib>=3.8.4
folium>=0.15.1
geopandas>=0.14.3